requirements.txt
aiogram
aiosqlite
python-dotenv
rule34Py
openai